requirements.txt
ansicon==1.89.0
blessed==1.20.0
blinker==1.7.0
click==8.1.7
colorama==0.4.6
editor==1.6.6
Flask==3.0.2
inquirer==3.2.4
itsdangerous==2.1.2
Jinja2==3.1.3
jinxed==1.2.1
MarkupSafe==2.1.5
pydobot==1.3.2
pyserial==3.4
readchar==4.0.6
runs==1.2.2
six==1.16.0
termcolor==2.4.0
tinydb==4.8.0
wcwidth==0.2.13
Werkzeug==3.0.2
xmod==1.8.1
yaspin==3.0.1
